+++ src/main/java/org/apache/skywalking/banyandb/v1/client/Value.java
@@ -22,6 +22,7 @@ import java.util.List;
 
 import com.google.common.base.Strings;
 import com.google.protobuf.ByteString;
+import com.google.protobuf.Timestamp;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import org.apache.skywalking.banyandb.model.v1.BanyandbModel;
@@ -129,6 +130,31 @@ public abstract class Value<T> {
     }
 
     /**
+     * The value of a timestamp type tag.
+     */
+    public static class TimestampTagValue extends Value<Long> implements Serializable<BanyandbModel.TagValue> {
+        private TimestampTagValue(Long value) {
+            super(value);
+        }
+
+        @Override
+        public BanyandbModel.TagValue serialize() {
+            Timestamp timestamp = toProtobufTimestamp(value);
+            return BanyandbModel.TagValue.newBuilder().setTimestamp(timestamp).build();
+        }
+    }
+
+    /**
+     * Utility method to convert milliseconds to protobuf Timestamp.
+     */
+    private static Timestamp toProtobufTimestamp(long millis) {
+        return Timestamp.newBuilder()
+                .setSeconds(millis / 1000)
+                .setNanos((int) ((millis % 1000) * 1000000))
+                .build();
+    }
+
+    /**
      * Construct a string tag
      *
      * @param val payload
@@ -191,6 +217,16 @@ public abstract class Value<T> {
             return nullTagValue();
         }
         return new LongArrayTagValue(val);
+    }
+
+    /**
+     * Construct a timestamp tag
+     *
+     * @param epochMilli payload in milliseconds
+     * @return Anonymous tag with timestamp payload
+     */
+    public static Serializable<BanyandbModel.TagValue> timestampTagValue(long epochMilli) {
+        return new TimestampTagValue(epochMilli);
     }
 
     public static Serializable<BanyandbModel.TagValue> nullTagValue() {